feat: add a feature for `Repository#update` to keep timestamp unchanged
- add the third argument (with default value) to `Repository#update`
- refactor tests for `FileSystemRepository` class
- bump version 0.5.3 -> 0.5.4
