+++ lib/textrepo/version.rb
@@ -1,3 +1,3 @@
 module Textrepo
-  VERSION = '0.5.3'
+  VERSION = '0.5.4'
 end
